dockerezing the services

+++ accounts/src/main/java/com/bank/accounts/repository/AccountsRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AccountsRepository extends JpaRepository<Accounts, Long> {
 
-    Accounts findByCustomerId(int customerId);
+    Accounts findByCustomerId(long customerId);
 }
